requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
opencv-python==4.9.0.80
numpy==1.26.4
pillow==10.2.0
torch==2.1.2
torchvision==0.16.2

realesrgan
basicsr
